[TQS-61] Reorganization, fixed Cucumber tests
