+++ Qourier/qourier_app/src/main/resources/templates/profile_rider.html
@@ -24,10 +24,10 @@
             <div class="col-12">
                 <h3>Statistics</h3>
 
-                <h5>Deliveries done over time</h5>
+                <h5 id="statistics-deliveries-over-time">Deliveries done over time</h5>
                 <img src="../static/images/grafana_placeholder.png" alt="Grafana placeholder for deliveries done over time" />
 
-                <h5>Average time spent on each delivery</h5>
+                <h5 id="statistics-average-time-spent">Average time spent on each delivery</h5>
                 <img src="../static/images/grafana_placeholder.png" alt="Grafana placeholder for average time spent" />
             </div>
         </div>
